Backend logic fix, bug fix on map editor

+++ marketplates-back/src/controllers/MapController.ts
@@ -3,8 +3,9 @@ import MapsModel from "../models/Maps.js";
 import jwt from "jsonwebtoken"
 import mongoose, { Types } from "mongoose";
 import PlaceIterationsModel from "../models/PlaceIterations.js";
-import { IMaps, PrivacyStatus, UserPrivileges } from "../types/mapTypes.js";
-import { IPlaceIteration, IPlaceUpdated } from "../types/placeIterationTypes.js";
+import { IMaps, PrivacyStatus, UserPrivileges } from "../common/types/mapTypes.js";
+import { IPlaceIteration, IPlaceUpdated } from "../common/types/placeIterationTypes.js";
+import checkOwnership from "../common/functions/checkOwnership.js";
 
 /**
    * Creates a map
@@ -71,12 +72,12 @@ export async function createMap(req, res) {
 
         await MapsModel.create(map);
 
-        res.status(201).json({
+        return res.status(201).json({
             message: '(201 Created)-Map successfully created',
             success: true
         });
     } catch (err) {
-        res.status(403).json({
+        return res.status(403).json({
             message: '(403 Forbidden)-The data sent created a map-type conflict',
             success: false
         });
@@ -96,13 +97,13 @@ export async function createMap(req, res) {
 export async function getAllMaps(req, res) {
     try {
         const allMaps = await MapsModel.find();
-        res.status(200).json({
+        return res.status(200).json({
             data: allMaps,
             message: '(200 OK)-Successfully fetched all maps',
             success: true
         });
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-No map was found',
             success: false
         });
@@ -122,13 +123,13 @@ export async function getAllMaps(req, res) {
 export async function getAllPublicMaps(req, res) {
     try {
         const allPublicMaps = await MapsModel.find({ privacyStatus: PrivacyStatus.Public });
-        res.status(200).json({
+        return res.status(200).json({
             data: allPublicMaps,
             message: '(200 OK)-Successfully fetched all official maps',
             success: true
         });
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-No map was found',
             success: false
         });
@@ -148,13 +149,13 @@ export async function getAllPublicMaps(req, res) {
 export async function getAllMapsAvailable(req, res) {
     try {
         const allAvailableMaps = await MapsModel.find({ $or: [{ privacyStatus: PrivacyStatus.Public }, { ownerId: req.params.ids }, { "participants.userId": req.params.ids }] });
-        res.status(200).json({
+        return res.status(200).json({
             data: allAvailableMaps,
             message: '(200 OK)-Successfully fetched all maps for the user',
             success: true
         });
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-No map was found',
             success: false
         });
@@ -180,13 +181,13 @@ export async function getUserMaps(req, res) {
         // Get the token's contents, verifying its validity in the process
         const decryptedCookie = jwt.verify(cookieValue, LOG_TOKEN_KEY);
         const allUserMaps = await MapsModel.find({ ownerId: decryptedCookie.userId });
-        res.status(200).json({
+        return res.status(200).json({
             data: allUserMaps,
             message: '(200 OK)-Successfully fetched all maps for the user',
             success: true
         });
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-No map was found',
             success: false
         });
@@ -207,13 +208,13 @@ export async function getMapsByIds(req, res) {
     try {
         // Ids, when sent in groups are sent in a single string, each Id tied to the others by a & character, hence the need for a split on that character
         const allPublicMaps = await MapsModel.find({ _id: { $in: req.params.ids.split("&") } });
-        res.status(200).json({
+        return res.status(200).json({
             data: allPublicMaps,
             message: '(200 OK)-Successfully fetched all maps by id',
             success: true
         });
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-No map was found',
             success: false
         });
@@ -243,81 +244,88 @@ export async function updateMapById(req, res) {
         const mapById: IMaps = await MapsModel.findOne({ _id: { $in: req.body.mapId } });
 
         if (!mapById) {
-            res.status(400).json({
+            return res.status(400).json({
                 message: '(404 Not Found)-The map to update was not found',
                 success: false
             });
         }
 
-        const iterations: IPlaceUpdated[] = req.body.formData.placeIterations;
-        const listOfIterationIds: Types.ObjectId[] = [];
+        if (checkOwnership(mapById.ownerId, decryptedCookie.userId, decryptedCookie.status)) {
 
-        for (const iteration of iterations) {
-            // Check if the place iteration already exists
-            const foundPreexistingIteration = await PlaceIterationsModel.find({ _id: { $in: [iteration._id] } })
+            const iterations: IPlaceUpdated[] = req.body.formData.placeIterations;
+            const listOfIterationIds: Types.ObjectId[] = [];
 
-            // If it does, we edit it with the new given values
-            if (foundPreexistingIteration) {
-                try {
-                    await PlaceIterationsModel.updateOne({ _id: { $in: foundPreexistingIteration[0]._id } }, {
-                        customName: iteration.name,
-                        customDescription: iteration.description,
-                        customTagIds: iteration.tagsList.map(tag => tag._id),
-                    })
-                    listOfIterationIds.push(foundPreexistingIteration[0]._id);
-                } catch {
-                    res.status(500).json({
-                        message: '(500 Internal Server Error)-A server side error has occured.',
-                        success: false
-                    });
-                }
-                // If it doesn't, we create it instead
-            } else {
-                try {
-                    const newId = new mongoose.Types.ObjectId();
-                    const iterationToCreate: IPlaceIteration = {
-                        _id: newId,
-                        associatedMapIds: [req.body.mapId],
-                        creatorId: decryptedCookie.userId,
-                        customName: iteration.name,
-                        customDescription: iteration.description,
-                        customTagIds: iteration.tagsList.map(tag => tag._id),
-                        gpsCoordinates: {
-                            longitude: iteration.gpsCoordinates.longitude,
-                            latitude: iteration.gpsCoordinates.latitude,
-                        },
-                        placeId: iteration._id,
+            for (const iteration of iterations) {
+                // Check if the place iteration already exists
+                const foundPreexistingIteration = await PlaceIterationsModel.find({ _id: { $in: [iteration._id] } })
 
+                // If it does, we edit it with the new given values
+                if (foundPreexistingIteration) {
+                    try {
+                        await PlaceIterationsModel.updateOne({ _id: { $in: foundPreexistingIteration[0]._id } }, {
+                            customName: iteration.name,
+                            customDescription: iteration.description,
+                            customTagIds: iteration.tagsList.map(tag => tag._id),
+                        })
+                        listOfIterationIds.push(foundPreexistingIteration[0]._id);
+                    } catch {
+                        return res.status(500).json({
+                            message: '(500 Internal Server Error)-A server side error has occured.',
+                            success: false
+                        });
                     }
-                    listOfIterationIds.push(newId);
+                    // If it doesn't, we create it instead
+                } else {
+                    try {
+                        const newId = new mongoose.Types.ObjectId();
+                        const iterationToCreate: IPlaceIteration = {
+                            _id: newId,
+                            associatedMapIds: [req.body.mapId],
+                            creatorId: decryptedCookie.userId,
+                            customName: iteration.name,
+                            customDescription: iteration.description,
+                            customTagIds: iteration.tagsList.map(tag => tag._id),
+                            gpsCoordinates: {
+                                longitude: iteration.gpsCoordinates.longitude,
+                                latitude: iteration.gpsCoordinates.latitude,
+                            },
+                            placeId: iteration._id,
 
-                    await PlaceIterationsModel.create(iterationToCreate);
-                } catch {
-                    res.status(500).json({
-                        message: '(500 Internal Server Error)-A server side error has occured.',
-                        success: false
-                    });
+                        }
+                        listOfIterationIds.push(newId);
+
+                        await PlaceIterationsModel.create(iterationToCreate);
+                    } catch {
+                        return res.status(500).json({
+                            message: '(500 Internal Server Error)-A server side error has occured.',
+                            success: false
+                        });
+                    }
                 }
             }
+
+            // Updating the map's data as well when changes are provided
+            await MapsModel.updateOne({ _id: { $in: req.body.mapId } }, {
+                description: req.body.formData.description ? sanitizeHtml(req.body.formData.description, { allowedTags: [] }) : mapById.description,
+                name: req.body.formData.name ? sanitizeHtml(req.body.formData.name, { allowedTags: [] }) : mapById.name,
+                participants: req.body.formData.participants,
+                privacyStatus: req.body.formData.privacyStatus,
+                placeIterationIds: listOfIterationIds,
+
+            })
+
+            return res.status(204).json({
+                message: '(204 No Content)-Map successfully updated',
+                success: true
+            });
+        } else {
+            return res.status(403).json({
+                message: '(403 Forbidden)-The user is not the owner of the map, or an admin and thus cannot update its data',
+                success: false
+            });
         }
-
-        // Updating the map's data as well when changes are provided
-        await MapsModel.updateOne({ _id: { $in: req.body.mapId } }, {
-            description: req.body.formData.description ? sanitizeHtml(req.body.formData.description, { allowedTags: [] }) : mapById.description,
-            name: req.body.formData.name ? sanitizeHtml(req.body.formData.name, { allowedTags: [] }) : mapById.name,
-            participants: req.body.formData.participants,
-            privacyStatus: req.body.formData.privacyStatus,
-            placeIterationIds: listOfIterationIds,
-
-        })
-
-        res.status(204).json({
-            message: '(204 No Content)-Map successfully updated',
-            success: true
-        });
-
     } catch (err) {
-        res.status(500).json({
+        return res.status(500).json({
             message: '(500 Internal Server Error)-A server side error has occured.',
             success: false
         });
@@ -336,15 +344,32 @@ export async function updateMapById(req, res) {
 */
 export async function deleteMapById(req, res) {
     try {
-        await MapsModel.deleteOne({ _id: { $in: req.body.mapId } });
+        // Get access token from the front end and the key that serves to create and verify them
+        const cookieValue = req.cookies.token;
+        const { LOG_TOKEN_KEY } = process.env;
 
-        res.status(204).json({
-            message: '(204 No Content)-Map successfully deleted',
-            success: true
-        });
+        // Get the token's contents, verifying its validity in the process
+        const decryptedCookie = jwt.verify(cookieValue, LOG_TOKEN_KEY);
 
+        // Find the matching map
+        const mapById: IMaps = await MapsModel.findOne({ _id: { $in: req.body.mapId } });
+
+
+        if (checkOwnership(mapById.ownerId, decryptedCookie.userId, decryptedCookie.status)) {
+            await MapsModel.deleteOne({ _id: { $in: req.body.mapId } });
+
+            return res.status(204).json({
+                message: '(204 No Content)-Map successfully deleted',
+                success: true
+            });
+        } else {
+            return res.status(403).json({
+                message: '(403 Forbidden)-The user is not the owner of the map, or an admin and thus cannot delete it',
+                success: false
+            });
+        }
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-Map to be deleted was not found',
             success: false
         });
@@ -363,15 +388,34 @@ export async function deleteMapById(req, res) {
 */
 export async function deleteMapsByIds(req, res) {
     try {
+        // Finding the matching maps
+        const mapsByIds: IMaps[] = await MapsModel.find({ _id: { $in: req.body.mapIds } });
+
+        // Get access token from the front end and the key that serves to create and verify them
+        const cookieValue = req.cookies.token;
+        const { LOG_TOKEN_KEY } = process.env;
+
+        // Get the token's contents, verifying its validity in the process
+        const decryptedCookie = jwt.verify(cookieValue, LOG_TOKEN_KEY);
+
+        for (const map of mapsByIds) {
+            if (!checkOwnership(map.ownerId, decryptedCookie.userId, decryptedCookie.status)) {
+                return res.status(403).json({
+                    message: '(403 Forbidden)-One of maps to be deleted does not belong to the user.',
+                    success: false
+                });
+            }
+        }
+
         await MapsModel.deleteMany({ _id: { $in: req.body.mapIds } });
 
-        res.status(204).json({
+        return res.status(204).json({
             message: '(204 No Content)-Maps successfully deleted',
             success: true
         });
 
     } catch (err) {
-        res.status(404).json({
+        return res.status(404).json({
             message: '(404 Not found)-Maps or several places to be deleted were not found',
             success: false
         });
